vinyl-store-html/backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
PyMySQL==1.1.0
